Namespace all Uno classes with UNO
