+++ lib/uno/constants.rb
@@ -1,5 +1,6 @@
-
-module Constants
-  MAX_PLAYERS = 4.freeze
-  MAX_CARDS   = 7.freeze
+module UNO
+  module Constants
+    MAX_PLAYERS = 4.freeze
+    MAX_CARDS   = 7.freeze
+  end
 end
