+++ lib/uno/player_factory.rb
@@ -1,5 +1,7 @@
-class PlayerFactory
-  def player(name, cards)
-    Player.new(name, cards)
+module UNO
+  class PlayerFactory
+    def player(name, cards)
+      Player.new(name, cards)
+    end
   end
 end
