+++ lib/uno/server.rb
@@ -1,0 +1,67 @@
+module UNO
+  class Server
+
+    attr_reader :deck, :pool, :players, :factory
+
+    def initialize(factory)
+      @deck    = Deck.buildDeck
+      @factory = factory
+      reset
+    end
+
+    def join_game? name
+      player_joined = can_add_more?
+      add(name) if player_joined
+      player_joined
+    end
+
+    def deal?
+      can_deal = has_players?
+      deal_cards if can_deal
+      can_deal
+    end
+
+    def see_cards_of name
+      player = players.find { |player| player.name == name }
+      return [] if player.nil?
+      player.cards.dup
+    end
+
+    def flip_top_card
+      pool.unshift(pool.pop)
+    end
+
+    def top_card
+      flip_top_card
+      pool.first
+    end
+
+    def reset
+      @pool    = deck.dup
+      @players = []
+    end
+
+    private
+
+    def can_add_more?
+      players.size < UNO::Constants::MAX_PLAYERS
+    end
+
+    def add(name)
+      players.push(new_player(name, []))
+    end
+
+    def new_player(name, cards)
+      factory.player(name, cards)
+    end
+
+    def has_players?
+      players.size > 0
+    end
+
+    def deal_cards
+      @pool = pool.shuffle
+      players.each { |player| player.cards = pool.pop(Constants::MAX_CARDS) }
+    end
+  end
+end
