+++ spec/uno/player_factory_spec.rb
@@ -1,6 +1,6 @@
-describe "Player factory" do
+describe "UNO::Playerfactory" do
   it "creates a player" do
-    jane = PlayerFactory.new.player("Jane", ["uno", "dos"])
+    jane = UNO::PlayerFactory.new.player("Jane", ["uno", "dos"])
     expect(jane.name).to eq("Jane")
     expect(jane.cards).to eq(["uno", "dos"])
   end
